refactor(area-chart): added useChartGrid
