+++ packages/components/charts/src/use-chart-reference-line.ts
@@ -62,7 +62,7 @@ export const useChartReferenceLine = ({
         ...omitObject(reChartsProps, ["label"]),
       }
     },
-    [styleClassName, theme],
+    [propList, styleClassName],
   )
 
   return { getReferenceLineProps }
